Run kresd tests with and without query name minimization
kresd_run.sh uses the new ability to read default value for NO_MINIMIZE
from environment variable to the whole run test suite twice:
With enabled and disabled query name minimization.

Tests which explicitly specify query-minimization option are not
affected and thus executed twice with the same options.
Practically this affects only 4 tests so I do not think it is worth
optimizing it.