Deckard: Use default Python version from system
Linux distributions are moving towards Python 3. As a consequence,
staying on Python 2 is causing problems. E.g. Deckard errored out
on missing 'dns' package but 'pip install' claimed that 'dns' is already
installed...

From now the Deckard is using default Python version so it should be
less surprising. If necessary, the default version can be overriden
using PYTHON variable in Makefile.

An important side-effect is that Deckard running under Python 3 is
easier to debug because Python 3 is able to report whole chains
of exceptions properly.