resolver: fix iter_lame_root
Root server address specification was missing so kresd was sending
queries to a wrong address. It worked because Deckard replies with
SERVFAIL for queries without defined answer.

The same situation was with enabled qname minimization. It worked
because of the default.

Now the answers are explicitly defined and Deckard does not log error
message about undefined answers.