kresd: run in non-interactive mode
For unknown reason, Deckard test might fail with "--- Logging error ---"
from Python interpreter. Experimentally I found out that this does not
happen anymore if kresd is executed with "-f 1", i.e. in non-interactive
mode. I do not know why, but this change seems to hide/solve the problem
and leads to more reliable test suite so it is now enabled by default.