explicit outgoing addresses to make PCAPs easier to read
Since migration to network namespaces PCAPs are harder to read because
source and destination addresses are always the same. This is AFAIK
caused by kernel routing and works as intended for real networks.

In our case all IP addresses are on local interfaces so we can
explicitly configure outgoing addresses in daemons and this will
again produce PCAPs where source address matches address assigned to
the daemon by Deckard.

At the moment Deckard assigns only one IPv4 or IPv6 address to binary
under test, i.e. this is not a complete fix: In cases where daemon has
IPv4 address assigned by Deckard but decides to contact IPv6 server it
will use the same address as source and destination.

Related: !169