iter_pcnamech.rpl: fix overlapping ranges
Previously, the scenario tested NOTHING.

So after the fix, it requires resolver to do parent-side query to get an
extra NS name which is needed in order to pass.

Knot Resolver is currently pretty parent centric, so fixed scenario will
fail. Let's reintroduce (or not) once we solve the parent/child problem
for good.